Added pickled logic

+++ growth_calculator.py
@@ -1,5 +1,6 @@
 """Calculate the growth for all the US stocks"""
 
+import os
 import pickle
 
 from ggm_calculator import InferParameters as Ip
@@ -12,6 +13,10 @@ class CalculateG:
 
     def __init__(self):
         """Initialize the class with the given routines"""
+        self.__pickled_file_path = os.path.join(os.getcwd(), "ticker_data.pickled")
+        if os.path.exists(self.__pickled_file_path):
+            self.__growth_parameters = pickle.load(open(self.__pickled_file_path, "rb"))
+            return
         self.__tickers = GetTickers().get_downloaded_tickers()
         self.__growth_parameters = self.__calculate_g()
         pickle.dump(self.__growth_parameters, open("ticker_data.pickled", "wb"))
@@ -26,7 +31,6 @@ class CalculateG:
         for letter in self.__tickers:
             for ticker in self.__tickers[letter]:
                 growth_parameters[ticker] = Ip(ticker).get_inferred_g()
-                print(growth_parameters[ticker])
         return growth_parameters
 
     def get_growth_rates(self):
